+++ src/main/java/bl4ckscor3/mod/globalxp/tileentity/XPBlockTileEntity.java
@@ -81,6 +81,8 @@ public class XPBlockTileEntity extends TileEntity implements ITickableTileEntity
 	{
 		storedXP = xp;
 		storedLevels = XPUtils.calculateStoredLevels(storedXP);
+		markDirty();
+		world.notifyBlockUpdate(pos, world.getBlockState(pos), world.getBlockState(pos), 2);
 	}
 
 	/**